Rename UsersList to UserList, UsersTypes to UserTypes in exercise 1
